+++ frontend/src/index.js
@@ -6,18 +6,18 @@ import {
 } from 'react-router-dom';
 import './index.css';
 import reportWebVitals from './reportWebVitals';
-import Login from './routes/login';
+import Login from './routes/Authentication/login';
 import Main from './routes/main';
 import Dashboard from './routes/dashboard';
 import Users from './routes/users';
-import IncomeStatement from './routes/incomeStatement';
-import BalanceSheet from './routes/balanceSheet';
-import CashflowStatement from './routes/cashflowStatements';
-import ChangesInEquity from './routes/changesInEquity';
-import GeneralLedger from './routes/generalLedger';
-import TrialBalance from './routes/trialBalance';
-import Deposits from './routes/deposits';
-import Collections from './routes/collections';
+import IncomeStatement from './routes/Accounting/incomeStatement';
+import BalanceSheet from './routes/Accounting/balanceSheet';
+import CashflowStatement from './routes/Accounting/cashflowStatements';
+import ChangesInEquity from './routes/Accounting/changesInEquity';
+import GeneralLedger from './routes/Accounting/generalLedger';
+import TrialBalance from './routes/Accounting/trialBalance';
+import Deposits from './routes/Accounting/deposits';
+import Collections from './routes/Accounting/collections';
 import { Provider } from 'react-redux';
 import store from './store';
